# packages/api-client/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MonitoramentoBuscaCreate": {
        "description": "Monitorar FUTURAS propostas de um município (opcionalmente por área/fonte).",
        "properties": {
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "canais": {
            "default": [
              "painel"
            ],
            "items": {
              "type": "string"
            },
            "title": "Canais",
            "type": "array"
          },
          "fonte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fonte"
          },
          "municipio_ibge": {
            "maxLength": 7,
            "minLength": 7,
            "title": "Municipio Ibge",
            "type": "string"
          }
        },
        "required": [
          "municipio_ibge"
        ],
        "title": "MonitoramentoBuscaCreate",
        "type": "object"
      },
      "MonitoramentoBuscaRead": {
        "properties": {
          "area": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Area"
          },
          "ativo": {
            "title": "Ativo",
            "type": "boolean"
          },
          "canais": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canais"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "fonte": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fonte"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "municipio_ibge": {
            "title": "Municipio Ibge",
            "type": "string"
          },
          "ultimo_alerta_em": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Ultimo Alerta Em"
          }
        },
        "required": [
          "id",
          "municipio_ibge",
          "ativo",
          "created_at"
        ],
        "title": "MonitoramentoBuscaRead",
        "type": "object"
      },
      "MonitoramentoCreate": {
        "properties": {
          "canais": {
            "default": [
              "painel"
            ],
            "items": {
              "type": "string"
            },
            "title": "Canais",
            "type": "array"
          },
          "proposta_id": {
            "format": "uuid",
            "title": "Proposta Id",
            "type": "string"
          }
        },
        "required": [
          "proposta_id"
        ],
        "title": "MonitoramentoCreate",
        "type": "object"
      },
      "MonitoramentoRead": {
        "properties": {
          "ativo": {
            "title": "Ativo",
            "type": "boolean"
          },
          "canais": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canais"
          },
          "created_at": {
            "format": "date-time",
            "title": "Created At",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "proposta_id": {
            "format": "uuid",
            "title": "Proposta Id",
            "type": "string"
          }
        },
        "required": [
          "id",
          "proposta_id",
          "ativo",
          "created_at"
        ],
        "title": "MonitoramentoRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "api/v1/auth/login"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "API pública v1 — propostas/repasses do governo (cache-first, RLS por usuário).",
    "title": "Hub Capture API",
    "version": "0.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/monitors": {
      "get": {
        "operationId": "listar_monitoramentos_api_v1_monitors_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoRead"
                  },
                  "title": "Response Listar Monitoramentos Api V1 Monitors Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Monitoramentos",
        "tags": [
          "monitoramentos"
        ]
      },
      "post": {
        "operationId": "criar_monitoramento_api_v1_monitors_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoramentoCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoramentoRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Criar Monitoramento",
        "tags": [
          "monitoramentos"
        ]
      }
    },
    "/api/v1/monitors/searches": {
      "get": {
        "operationId": "listar_buscas_api_v1_monitors_searches_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MonitoramentoBuscaRead"
                  },
                  "title": "Response Listar Buscas Api V1 Monitors Searches Get",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Listar Buscas",
        "tags": [
          "monitoramentos"
        ]
      },
      "post": {
        "operationId": "criar_busca_api_v1_monitors_searches_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MonitoramentoBuscaCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MonitoramentoBuscaRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Criar Busca",
        "tags": [
          "monitoramentos"
        ]
      }
    },
    "/api/v1/monitors/searches/{busca_id}": {
      "delete": {
        "operationId": "remover_busca_api_v1_monitors_searches__busca_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "busca_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Busca Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remover Busca",
        "tags": [
          "monitoramentos"
        ]
      }
    },
    "/api/v1/monitors/{monitoramento_id}": {
      "delete": {
        "operationId": "remover_monitoramento_api_v1_monitors__monitoramento_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "monitoramento_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Monitoramento Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Remover Monitoramento",
        "tags": [
          "monitoramentos"
        ]
      }
    },
...
  }
}